requirements.txt
Django==3.2.2
django-crispy-forms==1.11.2
django-hitcount==1.3.3
django-phonenumber-field==5.1.0
djangorestframework==3.12.4
mysqlclient==2.0.3
phonenumbers==8.12.23
Pillow==8.2.0asgiref==3.3.4
Django==3.2.2
django-crispy-forms==1.11.2
django-environ==0.4.5
django-etc==1.3.0
django-filter==2.4.0
django-hitcount==1.3.3
django-phonenumber-field==5.1.0
djangorestframework==3.12.4
mysqlclient==2.0.3
phonenumbers==8.12.23
Pillow==8.2.0
pytz==2021.1
sqlparse==0.4.1
